Week_5/Week_5/requirements.txt
sqlalchemy==2.0.22
psycopg2-binary
# if the binary fails to install,comment it out, uncomment the next line, save and reinstall
# psycopg2==2.9.7
